+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+### Unreleased
+* Add additional fields for job status webhook notifications
+
 ### 6.5.1 / 2022-09-16
 * Add additional `Event` fields
 * Fix issue with `EventParticipant` not sending status on new event creation